+++ rollup.config.mjs
@@ -4,14 +4,13 @@ export default [
   {
     input: 'src/index.ts',
     output: {
-      format: 'umd',
-      name: 'showJSError',
+      format: 'iife',
       file: './dist/show-js-error.js',
     },
     plugins: [typescript({ tsconfig: './tsconfig.json' })]
   },
   {
-    input: 'src/index.esm.ts',
+    input: 'src/index.ts',
     output: {
       format: 'es',
       file: './dist/show-js-error.esm.js'