+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v4.0.0
+- Fixed package.json properties.
+- Injected CSS to JS file.
+- Removed default export.
+- Fixes for Safari.
+
 ## v3.0.0
 - Dropped support for old browsers.
 - Dropped default exports.